+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to
 - 💄(frontend) add filtering to left panel #475
 - ✨(frontend) new share modal ui #489
 - ✨(frontend) add favorite feature #515
+- 🏷️(backend) add content-type to uploaded files #552
 - ✨(frontend) export pdf docx front side #537
 - 📝(documentation) Documentation about self-hosted installation #530
 - ✨(helm) helm versioning #530